Bump versions and update image handling
- Updated assembly and package versions to 1.0.7 and 1.2.14.
- Upgraded `Microsoft.CodeAnalysis.CSharp.Scripting` to 4.14.0.
- Changed `filename` type in `RenameFileDialog.cs` to `nuint`.
- Updated `Hexa.NET.ImGui`, `System.Numerics.Vectors`, and `System.Text.Json` packages.
- Added `Microsoft.Bcl.Numerics` version 9.0.7.
- Changed image parameter types in `ImageHelper.cs` to `ImTextureRef`.
- Updated `WidgetIcon` struct in `WidgetStyle.cs` for consistency.
- Modified `ImGuiManager.cs` for improved error handling.
- Updated `Hexa.NET.ImPlot` and `Hexa.NET.ImGui.Backends` versions in `TestApp.csproj`.

+++ Hexa.NET.ImGui.Widgets/Dialogs/RenameFileDialog.cs
@@ -104,7 +104,7 @@
 
         protected override void DrawContent()
         {
-            if (ImGui.InputText("New name", ref filename, (ulong)maxPathLength))
+            if (ImGui.InputText("New name", ref filename, (nuint)maxPathLength))
             {
                 destinationPath = Path.Combine(currentDirectory, filename);
                 warnMessage = null;
